feat: restructure typography components
