+++ packages/components/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './layouts';
 export * from './primitives';
+export * from './typography';
 export * from './notifications';
