+++ packages/components/src/components/typography/TextPairing/TextPairing.tsx
