+++ packages/components/src/components/typography/TextPairing/index.ts
