+++ packages/components/src/components/typography/TextPairing/types.ts
