+++ packages/components/src/components/typography/index.ts
@@ -1,0 +1,2 @@
+export * from './Text';
+export * from './TextPairing';
